NAV-24505: Legger til Journalføringsbehandlingstype til RestFerdigstillOppgaveKnyttJournalpost, flytter også enumen til domene pakken
