+++ src/main/kotlin/no/nav/familie/ba/sak/ekstern/restDomene/RestFerdigstillOppgaveKnyttJournalpost.kt
@@ -1,7 +1,7 @@
 package no.nav.familie.ba.sak.ekstern.restDomene
 
+import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.Journalføringsbehandlingstype
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingKategori
-import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingType
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingUnderkategori
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import java.time.LocalDateTime
@@ -12,7 +12,7 @@ data class RestFerdigstillOppgaveKnyttJournalpost(
     val opprettOgKnyttTilNyBehandling: Boolean = false,
     val navIdent: String,
     val bruker: NavnOgIdent,
-    val nyBehandlingstype: BehandlingType,
+    val nyBehandlingstype: Journalføringsbehandlingstype,
     val nyBehandlingsårsak: BehandlingÅrsak,
     val kategori: BehandlingKategori?,
     val underkategori: BehandlingUnderkategori?,
