+++ src/test/testdata/kotlin/no/nav/familie/ba/sak/datagenerator/JournalpostGenerator.kt
@@ -4,6 +4,7 @@ import no.nav.familie.ba.sak.ekstern.restDomene.NavnOgIdent
 import no.nav.familie.ba.sak.ekstern.restDomene.RestJournalføring
 import no.nav.familie.ba.sak.ekstern.restDomene.RestJournalpostDokument
 import no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner.DEFAULT_JOURNALFØRENDE_ENHET
+import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.Journalføringsbehandlingstype
 import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.Sakstype
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingKategori
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingUnderkategori
@@ -126,7 +127,7 @@ fun lagMockRestJournalføring(bruker: NavnOgIdent): RestJournalføring =
                 ),
             ),
         navIdent = "09123",
-        nyBehandlingstype = RestJournalføring.Journalføringsbehandlingstype.FØRSTEGANGSBEHANDLING,
+        nyBehandlingstype = Journalføringsbehandlingstype.FØRSTEGANGSBEHANDLING,
         nyBehandlingsårsak = BehandlingÅrsak.SØKNAD,
         fagsakType = FagsakType.NORMAL,
     )
